Re-construct the project root folder by switching the position of rails and react.
We switched them so that the root folder could have a clean environment of rails.
And then we put everything that comes from react inside the folder called client.
And then we positioned the client folder inside the application folder of rails.
This way in the future when we host this on heroku, inside the root project folder, we could initiate npm and it won't confuse heroku with the other package json file we have from react app which is inside app/client.